:truck: move about data out of modal directory and fix e2e tests
About content is no longer rendered in a modal, so move it to
pages/AboutView/AboutData.tsx. Fix e2e tests for API-driven links.
